requirements.txt
aiogram==3.2.0
python-decouple==3.8